Refactor to use floats instead of u8s for pos

+++ src/main.rs
@@ -17,8 +17,8 @@ fn main() {
         // .insert_resource(Msaa { samples: 4 })
         .insert_resource(ClearColor(Color::rgb(0.4, 0.4, 0.4)))
         .insert_resource(WindowDescriptor {
-            width: (game_map.width + 2 * game_map.pad_x) as f32 * game_map.sprite_size,
-            height: (game_map.height + 2 * game_map.pad_y) as f32 * game_map.sprite_size,
+            width: (game_map.width + 2. * game_map.pad_x) * game_map.sprite_size,
+            height: (game_map.height + 2. * game_map.pad_y) * game_map.sprite_size,
             title: "Cloud Surfer".to_string(),
             ..Default::default()
         })
